Add token auth to namespace deletion api
This commit adds token authentication to the namespace deletion api, which
is needed before the release of v7.0.0. The current usage of this allows
one to run from their command line:
`pelican namespace delete --prefix <namespace-prefix> --privkey \
</path/to/origin/privkey> --pubkey "" --namespace-url=<ns-url:port>

When run, the client creates and signs a token using the origin's privkey.
That token is sent to the registry along with a request to delete the prefix
from the database, and if the token passes verification/validation at the
registry using its stored jwks for that prefix, the prefix is deleted.